newsapi kotlin: refactor connectivity check again
