+++ DatabindingWithNewsApiKotlin/app/src/main/java/com/enpassio/databindingwithnewsapikotlin/ui/ArticleListFragment.kt
@@ -30,7 +30,6 @@ class ArticleListFragment : Fragment(), NewsAdapter.ArticleClickListener{
 
     private lateinit var mAdapter: NewsAdapter
     private lateinit var binding: NewsListBinding
-    private lateinit var snackbar: Snackbar
 
     init {
         retainInstance = true
@@ -60,7 +59,7 @@ class ArticleListFragment : Fragment(), NewsAdapter.ArticleClickListener{
         binding.included.viewModel = mViewModel
 
         //Claim the list from the view model and observe the results
-        mViewModel.articleList.observe(this, Observer { articles ->
+        mViewModel.articleList?.observe(this, Observer { articles ->
             if (!articles.isNullOrEmpty()) {
                 /*When articles are received, pass the articles to the adapter
                 And change uiState to SUCCESS. This will hide the loading indicator and show the list.*/
@@ -79,7 +78,7 @@ class ArticleListFragment : Fragment(), NewsAdapter.ArticleClickListener{
 
     private fun showSnack() {
         //Show a snack bar for warning about the network connection and prompt user to try again with a button
-        snackbar = Snackbar
+        val snackbar = Snackbar
             .make(binding.mainContent, R.string.no_network_connection, Snackbar.LENGTH_INDEFINITE)
             /*If user will click "Retry", we'll check the connection again,
                 and fetch the news, if there is network this time. Otherwise, snack will be shown again.*/
