+++ DatabindingWithNewsApiKotlin/app/src/main/java/com/enpassio/databindingwithnewsapikotlin/viewmodel/MainViewModel.kt
@@ -4,11 +4,7 @@ import android.app.Application
 import android.arch.lifecycle.AndroidViewModel
 import android.arch.lifecycle.LiveData
 import android.arch.lifecycle.MutableLiveData
-import android.content.Intent
 import android.databinding.ObservableField
-import android.net.Uri
-import android.util.Log
-import android.view.View
 import com.enpassio.databindingwithnewsapikotlin.data.Article
 import com.enpassio.databindingwithnewsapikotlin.data.NewsRepository
 import com.enpassio.databindingwithnewsapikotlin.utils.UIState
@@ -22,8 +18,13 @@ class MainViewModel(application: Application) :
     changes are automatically reflected in UI. */
     val uiState = ObservableField<UIState>(UIState.LOADING)
 
-    //Get the data from the repository and assign it to articleList
-    var articleList: LiveData<List<Article>?> = NewsRepository.articles
+    /*If article list is not null, return it, otherwise claim the
+   data from the repository, assign it to articleList and start fetching*/
+    var articleList: LiveData<List<Article>?>? = null
+        get() = field ?: NewsRepository.articles.also {
+            field = it
+            checkConnectionAndStartLoading()
+        }
 
     //Chosen article will be set later when an item is selected from the list
     var chosenArticle: Article? = null
@@ -34,13 +35,7 @@ class MainViewModel(application: Application) :
         get() = _showSnack
 
     init {
-        /*We don't check connection and restart loading if data is already there.
-        You might ask if data could be already there when viewModel is just initialized
-        Answer is yes, if user quits the app from back button, activities and viewModels are destroyed,
-        but static instance of singleton repository survives.*/
-        if(articleList.value.isNullOrEmpty()){
-            checkConnectionAndStartLoading()
-        }
+        _showSnack.value = false
     }
 
     /*If there is internet connection, start fetching data from the internet,
@@ -53,36 +48,12 @@ class MainViewModel(application: Application) :
             NewsRepository.startFetching()
             _showSnack.value = false
         } else {
-            /*If there is no connection, set uiState to NETWORK_ERROR
-            This will show an error message*/
-            uiState.set(UIState.NETWORK_ERROR)
+            //If there is no connection, show a snack message to warn user
             _showSnack.value = true
-        }
-    }
-
-    fun openWebSite(view: View) {
-        val articleUrl = chosenArticle?.articleUrl
-        if (!articleUrl.isNullOrEmpty()) {
-            //Parse string to uri
-            var webUri: Uri? = null
-            try {
-                webUri = Uri.parse(articleUrl)
-            } catch (e: Exception) {
-                Log.e(TAG, e.toString())
-            }
-
-            //Send an implicit intent to open the article in the browser
-            val webIntent = Intent(Intent.ACTION_VIEW)
-            with(webIntent) {
-                flags = Intent.FLAG_ACTIVITY_NEW_TASK
-                data = webUri
-                resolveActivity(view.context.packageManager)?.let {
-                    view.context.startActivity(this)
-                }
-            }
-
-            if (webIntent.resolveActivity(view.context.packageManager) != null) {
-                view.context.startActivity(webIntent)
+            /*Unless there is some previously fetched data to show,
+            set uiState to NETWORK_ERROR This will show an error message*/
+            if(articleList?.value.isNullOrEmpty()){
+                uiState.set(UIState.NETWORK_ERROR)
             }
         }
     }
