+++ DatabindingWithNewsApiKotlin/app/src/main/res/layout/fragment_details.xml
@@ -183,7 +183,6 @@
                         style="@style/Widget.Button.TextButton"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
-                        android:onClick="@{viewModel::openWebSite}"
                         android:text="@string/read_more"
                         app:backgroundTint="@color/colorPrimary"
                         app:layout_constraintBottom_toBottomOf="parent"
